+++ changelog.md
@@ -1,5 +1,6 @@
 * Fixed precise placement assist side num not updating after changing preset
 * Fixed game crashing on pressing delete in Settings Preset menu without choosing preset
+* Made GravChangerMode, MassChangerMode, PhysRopeMode, ScaleMode, SensorMode, SliderMode, ThrusterMode, presettable
 
 # 1.6.0
 * Added delete button to VEntityChanger